Made some review changes
